otel-demo-python/requirements.txt
Flask==3.0.3
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp==1.29.0
